+++ tests/Unit/CustomContext/CustomGraphQLController.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Rebing\GraphQL\Tests\Unit\CustomContext;
+
+use Illuminate\Http\Request;
+use Rebing\GraphQL\GraphQLController;
+
+class CustomGraphQLController extends GraphQLController
+{
+    protected function queryContext(string $query, ?array $params, string $schema, Request $request)
+    {
+        return new Context($request);
+    }
+}